renaming internal modules to avoid package namespace conflicts
