--- python/datamodel/git/__init__.py
@@ -1,1 +1,0 @@
-from .git import Git
