+++ python/datamodel/gitio/__init__.py
@@ -1,0 +1,1 @@
+from .gitio import Git
