+++ setup.cfg
@@ -47,6 +47,8 @@ scripts =
 	# vesitigial scripts - keeping for backwards compatibility
 	bin/datamodel_generate
 	bin/datamodel_install
+	bin/datamodel_tree
+	bin/datamodel_design
 
 [options.entry_points]
 console_scripts=
